+++ src/main/ml-modules/root/runDuringDeployment/generateRelatedListsConfig.mjs
@@ -26,7 +26,7 @@ const RELATION_KEYS_TO_SUPPRESS = [
   'classificationOfSet-containingItem-memberOf-usedForEvent',
   'classificationOfSet-usedForEvent',
 
-  //ignore all keys that start with publishedHere
+  //ignore all keys that start with publishedHere - #284
   'publishedHere-aboutAgent',
   'publishedHere-aboutConcept',
   'publishedHere-aboutPlace',
@@ -37,6 +37,16 @@ const RELATION_KEYS_TO_SUPPRESS = [
   'publishedHere-language',
   'publishedHere-publishedAt',
   'publishedHere-publishedBy',
+
+  //ignore all keys that end with publishedAt - #284
+  'subjectOfAgent-publishedAt',
+  'subjectOfConcept-publishedAt',
+  'subjectOfPlace-publishedAt',
+  'classificationOfWork-publishedAt',
+  'created-publishedAt',
+  'createdHere-publishedAt',
+  'languageOf-publishedAt',
+  'published-publishedAt',
 ];
 
 const targetId = TOKEN_RUNTIME_PARAM;